--- a/src/lib/segments.lua
+++ b/src/lib/segments.lua
@@ -18,7 +18,9 @@ function activateSegment(name,segment)
         s.onEnable(players[name])
     end
     players[name].activeSegments[segment]=true
-    _S.global.showMenu(name)
+    if s.menu then
+        _S.global.showMenu(name)
+    end
 end
 
 function deactivateSegment(name,segment)
@@ -54,7 +56,9 @@ function deactivateSegment(name,segment)
         s.onDisable(players[name])
     end
     players[name].activeSegments[segment]=nil
-    _S.global.showMenu(name)
+    if s.menu then
+        _S.global.showMenu(name)
+    end
 end
 
 function bindChatCommands()