--- a/src/segments/_global.lua
+++ b/src/segments/_global.lua
@@ -42,9 +42,6 @@ _S.global = {
         newGame=function()
             _S.global.tempMapName=nil
         end,
-        newPlayer=function(player)
-            _S.global.showMenu(player.name)
-        end,
         playerLeft=function(player)
             for n,p in pairs(players) do
                 if p.selected[player.name] then